Patch submitted by Paul Poley:
Thread's sleep(long), join() & join(long) can now be interrupted.

Also:
1) The Object on which a Thread is waiting is retained during interrupt so that
it is not possible to deallocate it if the wait times out before the interrupt
is able to process.
2) An issue is resolved so that if a notify is attempted while a wait thread is
waking up, a deadlock is not possible.

Include join() & join(long) in patch. This should finish up all methods related
to interrupt()